requirements.txt
click==8.1.7
Pillow==10.4.0
requests==2.32.3
numpy==2.0.1
matplotlib==3.9.0
mercantile==1.2.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
stripe==12.5.0
bcrypt==4.1.2
email-validator==2.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
